+++ rentacar/src/main/java/com/oguzhanturk/rentacar/business/request/car/CreateCarRequest.java
@@ -19,7 +19,7 @@ public class CreateCarRequest {
 
 	@NotNull(message = "Daily price should not be null")
 	@Positive(message = "Daily price should be positive")
-	private BigDecimal dailyPrice;
+	private double dailyPrice;
 
 	@Min(value = 1900, message = "Model year should not be least 1900 or null")
 	private int modelYear;
